feat(dashboard): revert gateway to direct when the Bankr key is removed
Extract syncGatewayProvider into lib/gateway.ts (shared by the auth route and
the secrets route) and keep gateway.provider tracking the BANKR_LLM_KEY secret
from every entry point:

- secrets DELETE of BANKR_LLM_KEY -> provider `direct` (classic)
- secrets POST   of BANKR_LLM_KEY -> provider `bankr`
- auth modal already handled bk_ -> bankr / else -> direct

So deleting the Bankr key no longer leaves provider=bankr pointing at a missing
secret (which fails the workflow); it falls back to the classic provider.
